server/requirements.txt
flask
flask-cors
requests
openai
groq
python-dotenv

llama_index
chromadb
llama-index-embeddings-jinaai